+++ samples/WebApplicationSample/Views/Shared/_Layout.cshtml
@@ -34,6 +34,9 @@
                     <li class="nav-item @Html.IsActive("WeChatPay")">
                         <a class="nav-link" asp-controller="WeChatPay" asp-action="Index">微信支付</a>
                     </li>
+                    <li class="nav-item @Html.IsActive("WeChatPayV3")">
+                        <a class="nav-link" asp-controller="WeChatPayV3" asp-action="Index">微信支付V3</a>
+                    </li>
                 </ul>
             </div>
         </div>
