+++ samples/WebApplicationSample/Views/WeChatPay/Index.cshtml
@@ -126,35 +126,5 @@
             <td><a href="https://pay.weixin.qq.com/wiki/doc/api/tools/mch_pay.php?chapter=24_7&index=4" target="_blank">https://fraud.mch.weixin.qq.com/risk/getpublickey</a></td>
             <td><a asp-controller="WeChatPay" asp-action="GetPublicKey">立即测试</a></td>
         </tr>
-        <tr>
-            <th scope="row">19</th>
-            <td>获取平台证书列表</td>
-            <td><a href="https://wechatpay-api.gitbook.io/wechatpay-api-v3/jie-kou-wen-dang/ping-tai-zheng-shu" target="_blank">https://api.mch.weixin.qq.com/v3/certificates</a></td>
-            <td><a asp-controller="WeChatPay" asp-action="GetCertificates">立即测试</a></td>
-        </tr>
-        <tr>
-            <th scope="row">20</th>
-            <td>APP支付-App下单</td>
-            <td><a href="https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pay/transactions/chapter3_1.shtml" target="_blank">https://api.mch.weixin.qq.com/v3/pay/transactions/app</a></td>
-            <td><a asp-controller="WeChatPay" asp-action="AppPayV3">立即测试</a></td>
-        </tr>
-        <tr>
-            <th scope="row">21</th>
-            <td>公众号支付-JSAPI下单</td>
-            <td><a href="https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pay/transactions/chapter3_2.shtml" target="_blank">https://api.mch.weixin.qq.com/v3/pay/transactions/jsapi</a></td>
-            <td><a asp-controller="WeChatPay" asp-action="PubPayV3">立即测试</a></td>
-        </tr>
-        <tr>
-            <th scope="row">22</th>
-            <td>扫码支付-Native下单</td>
-            <td><a href="https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pay/transactions/chapter3_3.shtml" target="_blank">https://api.mch.weixin.qq.com/v3/pay/transactions/native</a></td>
-            <td><a asp-controller="WeChatPay" asp-action="QrCodePayV3">立即测试</a></td>
-        </tr>
-        <tr>
-            <th scope="row">23</th>
-            <td>H5支付-H5下单</td>
-            <td><a href="https://pay.weixin.qq.com/wiki/doc/apiv3/wxpay/pay/transactions/chapter3_4.shtml" target="_blank">https://api.mch.weixin.qq.com/v3/pay/transactions/h5</a></td>
-            <td><a asp-controller="WeChatPay" asp-action="H5PayV3">立即测试</a></td>
-        </tr>
     </tbody>
 </table>
