+++ samples/WebApplicationSample/Views/WeChatPayV3/GetCertificates.cshtml
@@ -3,14 +3,14 @@
 }
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a asp-controller="WeChatPay" asp-action="Index">微信支付</a></li>
+        <li class="breadcrumb-item"><a asp-controller="WeChatPayV3" asp-action="Index">微信支付V3</a></li>
         <li class="breadcrumb-item active" aria-current="page">@ViewData["Title"]</li>
     </ol>
 </nav>
 <br />
 <div class="card">
     <div class="card-body">
-        <form asp-controller="WeChatPay" asp-action="GetCertificates">
+        <form asp-controller="WeChatPayV3" asp-action="GetCertificates">
             <button type="submit" class="btn btn-primary">提交请求</button>
         </form>
         <hr />
